project/requirements.txt
fastapi
pydantic
uvicorn
alembic
python-dotenv
sqlalchemy==1.4.35
asyncpg==0.26.0
pytz
httpx